agent: Move `Context` to state machine module
